Added shebang and main function.
So the shebang - first line where it says #!/usr/bin/env python3 - this tells linux what program it should be using to load the file - my understanding is that that gets ignored in windows but if it doesn't and that causes the script to not run properly in windows then just delete it.  Windows has the concept of 'file associations' eg: anything named *.py gets run by python, but linux doesn't and needs that hint at the beginning of the file.

Secondly, and more importantly, I created a main function.  This is important because now you can import this class into any other script you write in the future and you don't have to rewrite it or anything.  from another class you could do something like:

```
from pypets import Pet
```
and then you could just use that class as if it was defined in the script your importing it to - so you don't have to repeat your code and there's only one place you have to maintain it.

Some references:
shebang: https://stackoverflow.com/questions/7574453/shebang-notation-python-scripts-on-windows-and-linux
def main: https://stackoverflow.com/questions/419163/what-does-if-name-main-do